[0.5.2 All] Script approximate length in Minutes always displayed somewhere on the screen #193
If we use QTextDocument::pageCount() to estimate total time, then we
will always get rounded off to 1 min for both total time and current
time irrespective of the cursor position within the page. This issue has
now been fixed.

+++ src/document/screenplaytextdocument.h
@@ -14,11 +14,12 @@
 #ifndef SCREENPLAYTEXTDOCUMENT_H
 #define SCREENPLAYTEXTDOCUMENT_H
 
+#include <QTime>
+#include <QtMath>
 #include <QTextDocument>
 #include <QQmlParserStatus>
 #include <QPagedPaintDevice>
 #include <QAbstractTextDocumentLayout>
-#include <QTime>
 
 #include "scene.h"
 #include "formatting.h"
@@ -206,7 +207,7 @@ protected:
 private:
     void init();
     void setUpdating(bool val);
-    void setPageCount(int val);
+    void setPageCount(qreal val);
     void setCurrentPageAndPosition(int page, qreal pos);
     void resetFormatting();
     void resetTextDocument();
